+++ backend/account/urls.py
@@ -15,8 +15,6 @@ urlpatterns = [
     path("login/", views.login),
     path("logout/", views.logout),
     path("csrf-token/", views.get_csrf_token),
-    path("follow/", views.followView.as_view()),
-    path("follows/", views.get_follows),
     path("profile/nickname/", views.change_nickname),
     path("profile/image/", views.change_profile_image),
 ]
