+++ backend/follows/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class FollowsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "follows"
