+++ backend/transcendence/settings.py
@@ -43,6 +43,7 @@ INSTALLED_APPS = [
     "rest_framework_simplejwt",
     "corsheaders",
     "account",
+    "follows",
 ]
 
 REST_FRAMEWORK = {
